% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_code.R
\name{event_price_count_plot}
\alias{event_price_count_plot}
\title{Histogram of Event Prices}
\usage{
event_price_count_plot(events_df)
}
\arguments{
\item{events_df}{A data frame containing event data.}
}
\value{
A ggplot histogram of ticket prices.
}
\description{
Generates a histogram showing the distribution of average ticket prices.
}
\examples{
\dontrun{
event_price_count_plot(events_df)
}
}
