% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_code.R
\name{get_venue_info}
\alias{get_venue_info}
\title{Get Venue Information}
\usage{
get_venue_info(events_df)
}
\arguments{
\item{events_df}{A data frame containing event data.}
}
\value{
A list with the venue having the most and least events.
}
\description{
Extracts venue information and counts the number of events per venue.
}
\examples{
\dontrun{
venue_info <- get_venue_info(events_df)
print(venue_info)
}
}
