% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_code.R
\name{setup_ticketmaster_api_key}
\alias{setup_ticketmaster_api_key}
\title{Set Up API Key for Ticketmaster}
\usage{
setup_ticketmaster_api_key()
}
\description{
To use this package, you need to set your Ticketmaster API key as an environment variable.
}
\details{
The API key must be stored as an environment variable to be used automatically in functions that require it.
}
\examples{
# Set your API key (only needs to be done once per session)
Sys.setenv(TICKETMASTER_API_KEY = "your_api_key_here")

# Retrieve the API key (example usage)
api_key <- get_ticketmaster_api_key()
print(api_key)

# Now you can call functions without passing the API key each time
events_df <- get_full_ticketmaster_data()

}
