% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TicketmasterR-package.R
\docType{package}
\name{TicketmasterR-package}
\alias{TicketmasterR}
\alias{TicketmasterR-package}
\title{TicketmasterR: An R Wrapper for the Ticketmaster API}
\description{
The \code{TicketmasterR} package provides functions to interact with the
Ticketmaster API, allowing users to fetch and analyze event data,
retrieve ticket prices, and visualize event trends.
}
\details{
\subsection{Main Functions:}{
\itemize{
\item \code{get_full_ticketmaster_data()}: Fetch event data from Ticketmaster.
\item \code{ticketmaster_analysis()}: Analyze event trends, including highest and lowest ticket prices.
}
}

\subsection{Getting Started:}{

To start using the package, set your API key:

\if{html}{\out{<div class="sourceCode">}}\preformatted{sys.setenv(TICKETMASTER_API_KEY = "your_api_key_here")
}\if{html}{\out{</div>}}

Then, fetch event data:

\if{html}{\out{<div class="sourceCode">}}\preformatted{events_df <- get_full_ticketmaster_data()
}\if{html}{\out{</div>}}
}
}
\author{
\strong{Maintainer}: Aamna Amer - Jason Suwito - Kelsey Strachan \email{jasonns@student.ubc.ca}

}
\keyword{internal}
