% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TicketmasterR-package.R
\docType{package}
\name{TicketmasterR-package}
\alias{TicketmasterR}
\alias{TicketmasterR-package}
\title{TicketmasterR: An R Wrapper for the Ticketmaster API}
\description{
The \code{TicketmasterR} package provides functions to interact with the
Ticketmaster API, allowing users to fetch and analyze event data,
retrieve ticket prices, and visualize event trends.
}
\details{
\subsection{Main Functions:}{
\itemize{
\item \code{get_full_ticketmaster_data()}: Retrieves and processes event data from the Ticketmaster API.
\item \code{print_ticketmaster_analysis()}: Displays a summary of Ticketmaster event data, including venue and ticket price statistics.
\item \code{ticketmaster_analysis()}: Analyzes event data from Ticketmaster, including highest and lowest ticket prices and venue statistics.
\item \code{get_venue_info()}: Extracts venue information and counts the number of events per venue.
\item \code{get_highest_ticket_price()}: Identifies the event with the highest ticket price from the given event data.
\item \code{get_lowest_ticket_price()}: Identifies the event with the lowest ticket price from the given event data.
\item \code{event_class_plot()}: Generates a bar chart showing the number of events available per genre.
\item \code{avg_price_class_plot()}: Generates a bar chart displaying the average ticket price for events, grouped by genre.
\item \code{event_price_count_plot()}: Generates a histogram showing the distribution of average ticket prices.
\item \code{avg_event_price_line_plot()}: Generates a line plot showing the average ticket price trend for the last 200 events.
\item \code{event_hourly_distribution_plot()}: Generates a bar chart displaying the distribution of event counts across different hours of the day.
\item \code{event_day_distribution_plot()}: Generates a bar chart displaying the number of events occurring on each day of the week.
}
}

\subsection{Getting Started:}{

To start using the package, set your API key:

\if{html}{\out{<div class="sourceCode">}}\preformatted{sys.setenv(TICKETMASTER_API_KEY = "your_api_key_here")
}\if{html}{\out{</div>}}

Then, fetch event data:

\if{html}{\out{<div class="sourceCode">}}\preformatted{events_df <- get_full_ticketmaster_data()
}\if{html}{\out{</div>}}
}
}
\author{
\strong{Maintainer}: Aamna Amer - Jason Suwito - Kelsey Strachan \email{jasonns@student.ubc.ca}

}
\keyword{internal}
